move top level go packages into pkg/
