+++ BUILD.bazel
@@ -78,10 +78,8 @@ filegroup(
         "//config:all-srcs",
         "//def:all-srcs",
         "//experiment:all-srcs",
-        "//flagutil:all-srcs",
         "//gcsweb/cmd/gcsweb:all-srcs",
         "//gcsweb/pkg/version:all-srcs",
-        "//ghclient:all-srcs",
         "//ghproxy:all-srcs",
         "//gopherage:all-srcs",
         "//greenhouse:all-srcs",
@@ -96,6 +94,8 @@ filegroup(
         "//maintenance/aws-janitor:all-srcs",
         "//maintenance/migratestatus:all-srcs",
         "//metrics:all-srcs",
+        "//pkg/flagutil:all-srcs",
+        "//pkg/ghclient:all-srcs",
         "//prow:all-srcs",
         "//robots/commenter:all-srcs",
         "//robots/issue-creator:all-srcs",
