+++ pkg/README.md
@@ -1,0 +1,3 @@
+# pkg
+
+pkg/ contains common go packages used throughout test-infra
