+++ pkg/flagutil/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["flagutil.go"],
-    importpath = "k8s.io/test-infra/flagutil",
+    importpath = "k8s.io/test-infra/pkg/flagutil",
     visibility = ["//visibility:public"],
 )
