+++ prow/cmd/hook/BUILD.bazel
@@ -33,7 +33,7 @@ go_library(
     srcs = ["main.go"],
     importpath = "k8s.io/test-infra/prow/cmd/hook",
     deps = [
-        "//flagutil:go_default_library",
+        "//pkg/flagutil:go_default_library",
         "//prow/config:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/hook:go_default_library",
