+++ prow/cmd/tide/BUILD.bazel
@@ -13,7 +13,7 @@ go_library(
     importpath = "k8s.io/test-infra/prow/cmd/tide",
     visibility = ["//visibility:private"],
     deps = [
-        "//flagutil:go_default_library",
+        "//pkg/flagutil:go_default_library",
         "//prow/config:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/logrusutil:go_default_library",
