+++ prow/cmd/tide/main.go
@@ -28,7 +28,7 @@ import (
 
 	"github.com/sirupsen/logrus"
 
-	flagutil "k8s.io/test-infra/flagutil"
+	"k8s.io/test-infra/pkg/flagutil"
 	"k8s.io/test-infra/prow/config"
 	prowflagutil "k8s.io/test-infra/prow/flagutil"
 	"k8s.io/test-infra/prow/logrusutil"
