+++ prow/external-plugins/refresh/BUILD.bazel
@@ -10,7 +10,7 @@ go_library(
     importpath = "k8s.io/test-infra/prow/external-plugins/refresh",
     visibility = ["//visibility:private"],
     deps = [
-        "//flagutil:go_default_library",
+        "//pkg/flagutil:go_default_library",
         "//prow/config:go_default_library",
         "//prow/flagutil:go_default_library",
         "//prow/github:go_default_library",
