+++ prow/flagutil/BUILD.bazel
@@ -37,5 +37,5 @@ go_test(
     name = "go_default_test",
     srcs = ["kubernetes_test.go"],
     embed = [":go_default_library"],
-    deps = ["//flagutil:go_default_library"],
+    deps = ["//pkg/flagutil:go_default_library"],
 )
