+++ prow/prstatus/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "k8s.io/test-infra/prow/prstatus",
     visibility = ["//visibility:public"],
     deps = [
-        "//ghclient:go_default_library",
+        "//pkg/ghclient:go_default_library",
         "//prow/config:go_default_library",
         "//prow/github:go_default_library",
         "//vendor/github.com/google/go-github/github:go_default_library",
@@ -35,7 +35,7 @@ go_test(
     srcs = ["prstatus_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//ghclient:go_default_library",
+        "//pkg/ghclient:go_default_library",
         "//prow/config:go_default_library",
         "//prow/github:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
