+++ robots/issue-creator/creator/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "k8s.io/test-infra/robots/issue-creator/creator",
     visibility = ["//visibility:public"],
     deps = [
-        "//ghclient:go_default_library",
+        "//pkg/ghclient:go_default_library",
         "//robots/issue-creator/testowner:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/github.com/google/go-github/github:go_default_library",
